requirements.txt
Django==3.2.9
psycopg2==2.9.2
djangorestframework==3.12.4
